Modified Files:  	Utility.py  ---------------------------------------------------------------------- Added a runtime patch for minidom._clone_node.   In WSDL _import, an imported document's definition node is not cloned, but its children are.
A definition node's attributes are handed to its cloned children, this
problem surfaced while cloning a schema node that declared an xmlns prefixed
attribute name.

This is the appropriate place to handle this problem since namespace
declarations can appear in any start-tag.


I noticed that the PullDom patches in Utility are included in python2.2
and PyXML-0.8.1, they are not in python2.1


 ----------------------------------------------------------------------


git-svn-id: http://svn.code.sf.net/p/pywebsvcs/code/trunk@164 7150bf37-e60d-0410-b93f-83e91ef0e581