in _Call.__call__(), don't pass TC.Any(aslist=1) to self.binding.Receive().
self.binding.Receive() has an if statement that checks if the replytype
parameter is None.  if replytype is None, Receive() tries to locate the
typecode in the typesmodule based upon the response parameter.  with
__call__() setting repltype to TC.Any(aslist=1), the code was unable to
hit this if statement and begin the process of discovering the typecode.

there's a hook in Receive() that says that if the typecode cannot be
discovered, set it to TC.Any(aslist=1), thus having the same effect as
passing TC.Any(aslist=1) to Receive().


git-svn-id: http://svn.code.sf.net/p/pywebsvcs/code/trunk@620 7150bf37-e60d-0410-b93f-83e91ef0e581