modified code called when use_wsdl argument is passed to ServiceProxy constructor.
when no arguments were passed to an operation, the code failed to generate a
typecode.  i changed this by checking if the request is None, and if it is use
the Any typecode setting oname=Name where Name is the name of the operation
invoked.

also, a bug existed (at least in my environment) for optional elements in a
complex type.  the code was not detecting the minOccurs value properly.  when
it returned as u'0', the "not minOccurs > 0" test was failing.  wrapping the
minOccurs initialization in an int() fixed the problem.


git-svn-id: http://svn.code.sf.net/p/pywebsvcs/code/trunk@664 7150bf37-e60d-0410-b93f-83e91ef0e581