M    ZSI/TC.py
The bug I thought you came across is a known problem with minidom resolving the default namespace.

	""Minidom bug in ParsedSoap: Default namespace is empty - ID: 1810600"""

But I plugged in 4Suite, without any modifications from my formula ( maybe a versioning difference in regards to 4Suite ), and got the same error.  So I looked at TC.py and found the problem, it's trivial, probably occurred as a result of some refactoring since it is correctly handled elsewhere.


git-svn-id: http://svn.code.sf.net/p/pywebsvcs/code/trunk@1488 7150bf37-e60d-0410-b93f-83e91ef0e581